streamlining and cleanup
